# brev.ly-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Brev.ly API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/uploads/{id}": {
      "delete": {
        "summary": "Remove um link encurtado do banco de dados",
        "tags": [
          "Uploads"
        ],
        "description": "Deleta permanentemente um registro com base no ID informado.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID único do upload que será removido"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Confirmação de remoção bem-sucedida"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Upload não encontrado ou ID inválido"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/check/{slug}": {
      "get": {
        "summary": "Validação de link encurtado",
        "tags": [
          "Links"
        ],
        "description": "Verifica a existência de um slug e retorna a URL original se existir",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "exists"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "exists"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "exists"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": []
}